Add GroupID to Task struct and update MockTasks() function
This commit adds a new field, GroupID, to the Task struct in the scheduler package. It also updates the MockTasks() function in the scheduler package to assign a value to the GroupID field when creating tasks. This change allows for better organization and categorization of tasks based on their group.